+++ adapters/src/main/java/com/squareup/moshi/adapters/Rfc3339DateJsonAdapter.java
@@ -13,24 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.moshi;
+package com.squareup.moshi.adapters;
 
+import com.squareup.moshi.JsonAdapter;
+import com.squareup.moshi.JsonReader;
+import com.squareup.moshi.JsonWriter;
 import java.io.IOException;
 import java.util.Date;
 
 /**
- * @deprecated this class moved to avoid a package name conflict in the Java Platform Module System.
- *     The new class is {@code com.squareup.moshi.adapters.Rfc3339DateJsonAdapter}.
+ * Formats dates using <a href="https://www.ietf.org/rfc/rfc3339.txt">RFC 3339</a>, which is
+ * formatted like {@code 2015-09-26T18:23:50.250Z}.
  */
 public final class Rfc3339DateJsonAdapter extends JsonAdapter<Date> {
-  com.squareup.moshi.adapters.Rfc3339DateJsonAdapter delegate
-      = new com.squareup.moshi.adapters.Rfc3339DateJsonAdapter();
-
-  @Override public Date fromJson(JsonReader reader) throws IOException {
-    return delegate.fromJson(reader);
+  @Override public synchronized Date fromJson(JsonReader reader) throws IOException {
+    String string = reader.nextString();
+    return Iso8601Utils.parse(string);
   }
 
-  @Override public void toJson(JsonWriter writer, Date value) throws IOException {
-    delegate.toJson(writer, value);
+  @Override public synchronized void toJson(JsonWriter writer, Date value) throws IOException {
+    String string = Iso8601Utils.format(value);
+    writer.value(string);
   }
 }
