+++ examples/src/main/java/com/squareup/moshi/recipes/ReadAndWriteRfc3339Dates.java
@@ -17,7 +17,7 @@ package com.squareup.moshi.recipes;
 
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
-import com.squareup.moshi.Rfc3339DateJsonAdapter;
+import com.squareup.moshi.adapters.Rfc3339DateJsonAdapter;
 import com.squareup.moshi.recipes.models.Tournament;
 import java.util.Calendar;
 import java.util.Date;
